relocate header components and back button
fix back button imports
