+++ src/components/header/BaseHeader.tsx
@@ -1,10 +1,10 @@
 import type { ReactNode } from 'react';
 import { ProfileDropDown } from './ProfileDropDown';
 import { getServerAuthSession } from '@src/server/auth';
-import { ClubSearchBar } from './searchBar/ClubSearchBar';
-import { EventSearchBar } from './searchBar/EventSearchBar';
+import { ClubSearchBar } from '../searchBar/ClubSearchBar';
+import { EventSearchBar } from '../searchBar/EventSearchBar';
 import SignInButton from './signInButton';
-import MobileNav from './nav/MobileNav';
+import MobileNav from '../nav/MobileNav';
 import { api } from '@src/trpc/server';
 
 export const BaseHeader = async ({ children }: { children: ReactNode }) => {
